fan_service: use floats and avoid rounding inaccuracies

+++ fboss/platform/fan_service/ControlLogic.cpp
@@ -194,7 +194,7 @@ std::tuple<bool, int, uint64_t> ControlLogic::readFanRpm(const Fan& fan) {
 }
 
 void ControlLogic::updateTargetPwm(const Sensor& sensor) {
-  int16_t targetPwm{0};
+  float targetPwm{0};
   TempToPwmMap tableToUse;
   auto& readCache = sensorReadCaches_[*sensor.sensorName()];
   const auto& pwmCalcType = *sensor.pwmCalcType();
@@ -301,7 +301,7 @@ void ControlLogic::getOpticsUpdate() {
       continue;
     }
 
-    int aggOpticPwm = 0;
+    float aggOpticPwm = 0;
     const auto& aggregationType = *optic.aggregationType();
 
     if (aggregationType == constants::OPTIC_AGGREGATION_TYPE_MAX()) {
@@ -422,10 +422,7 @@ bool ControlLogic::isFanPresentInDevice(const Fan& fan) {
   return fanPresent;
 }
 
-bool ControlLogic::programFan(
-    const Zone& zone,
-    const Fan& fan,
-    int16_t fanPwm) {
+bool ControlLogic::programFan(const Zone& zone, const Fan& fan, float fanPwm) {
   bool writeSuccess{false};
 
   int pwmRawValue = static_cast<int>(
@@ -479,12 +476,12 @@ void ControlLogic::programLed(const Fan& fan, bool fanFailed) {
       fmt::format(kLedWriteFailure, *fan.fanName()), !ret);
 }
 
-int16_t ControlLogic::calculateZonePwm(const Zone& zone, bool boostMode) {
+float ControlLogic::calculateZonePwm(const Zone& zone, bool boostMode) {
   auto zoneType = *zone.zoneType();
-  int16_t zonePwm{0};
+  float zonePwm{0};
   int totalPwmConsidered{0};
   for (const auto& sensorName : *zone.sensorNames()) {
-    int16_t pwmForThisSensor;
+    float pwmForThisSensor;
     if (sensorReadCaches_.find(sensorName) != sensorReadCaches_.end()) {
       pwmForThisSensor = sensorReadCaches_[sensorName].targetPwmCache;
     } else if (opticReadCaches_.find(sensorName) != opticReadCaches_.end()) {
@@ -507,7 +504,7 @@ int16_t ControlLogic::calculateZonePwm(const Zone& zone, bool boostMode) {
     zonePwm /= totalPwmConsidered;
   }
   if (boostMode) {
-    zonePwm = std::max(zonePwm, *config_.pwmBoostValue());
+    zonePwm = std::max(zonePwm, (float)*config_.pwmBoostValue());
   }
 
   XLOG(INFO) << fmt::format(
@@ -531,7 +528,7 @@ void ControlLogic::setTransitionValue() {
                 *fan.fanName()) == zone.fanNames()->end()) {
           continue;
         }
-        int16_t fanPwm = *config_.pwmTransitionValue();
+        float fanPwm = *config_.pwmTransitionValue();
         bool fanFailed = programFan(zone, fan, fanPwm);
 
         for (auto& [key, pidLogic] : pidLogics_) {
@@ -656,7 +653,7 @@ void ControlLogic::updateControl(std::shared_ptr<SensorData> pS) {
   // STEP 5: Calculate and program fan PWMs
   fanStatuses_.withWLock([&](auto& fanStatuses) {
     for (const auto& zone : *config_.zones()) {
-      int16_t zonePwm = calculateZonePwm(zone, boostMode);
+      float zonePwm = calculateZonePwm(zone, boostMode);
       for (const auto& fan : *config_.fans()) {
         if (std::find(
                 zone.fanNames()->begin(),
@@ -665,7 +662,7 @@ void ControlLogic::updateControl(std::shared_ptr<SensorData> pS) {
           continue;
         }
 
-        int16_t fanPwm = calculateFanPwm(
+        float fanPwm = calculateFanPwm(
             *zone.slope(),
             *fanStatuses[*fan.fanName()].pwmToProgram(),
             zonePwm);
@@ -687,19 +684,19 @@ void ControlLogic::updateControl(std::shared_ptr<SensorData> pS) {
   });
 }
 
-int16_t ControlLogic::calculateFanPwm(
-    uint16_t slope,
-    int16_t currentFanPwm,
-    int16_t zonePwm) {
-  int16_t newFanPwm = zonePwm;
+float ControlLogic::calculateFanPwm(
+    float slope,
+    float currentFanPwm,
+    float zonePwm) {
+  float newFanPwm = zonePwm;
   if ((slope == 0) || (currentFanPwm == 0)) {
     newFanPwm = zonePwm;
   } else if (std::abs(currentFanPwm - zonePwm) > slope) {
     newFanPwm = currentFanPwm + ((zonePwm > currentFanPwm) ? slope : -slope);
   }
 
-  newFanPwm = std::min(newFanPwm, *config_.pwmUpperThreshold());
-  newFanPwm = std::max(newFanPwm, *config_.pwmLowerThreshold());
+  newFanPwm = std::min(newFanPwm, (float)*config_.pwmUpperThreshold());
+  newFanPwm = std::max(newFanPwm, (float)*config_.pwmLowerThreshold());
 
   std::optional<int> fanHoldPwm = fanHoldPwm_.load();
   if (fanHoldPwm.has_value()) {
@@ -709,7 +706,7 @@ int16_t ControlLogic::calculateFanPwm(
   return newFanPwm;
 }
 
-void ControlLogic::updatePwmState(const Zone& zone, int16_t fanPwm) {
+void ControlLogic::updatePwmState(const Zone& zone, float fanPwm) {
   for (const auto& sensorName : *zone.sensorNames()) {
     if (pidLogics_.find(sensorName) != pidLogics_.end()) {
       pidLogics_.at(sensorName)->updateLastPwm(fanPwm);
